chore: re-organize components
