+++ client/src/navbar.js
@@ -1,11 +1,9 @@
 import React from 'react';
-import SpotifyDisplay from './spotify-display';
 
 const Navbar = () => (
 	<div>
 		<nav>
 			<h1>MOODIFY</h1>
-			<SpotifyDisplay />
 		</nav>
 		<hr />
 	</div>
